4. Basic Queue Operations
4. Basic Queue Operations 
You will be given an integer N representing the number of elements to enqueue (add), an integer S representing the
number of elements to dequeue (remove/poll) from the queue, and finally, an integer X, an element that you
should check whether is present in the queue. If it is - prints true on the console, if it is not - print the smallest
element currently present in the queue.