1. Compare Matrices
1. Compare Matrices 
Write a program that reads two integer matrices (2D arrays) from the console and compares them element by
ment. For better code reusability, you could make the comparison in a method that returns true if they are
equal and false if not. 
Each matrix definition on the console will contain a line with a positive integer number R – the number of rows in
the matrix and C – the number of columns – followed by R lines containing the C numbers, separated by spaces
each line will have an equal amount of numbers. 
The matrices will have at most 10 rows and at most 10 columns.
rint "equal" if the matrices match and "not equal" if they don't match.